# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Condominios API",
        "description": "API para administración de condominios.\n\nCredenciales locales para probar en Swagger:\n- Email: swagger.admin@example.com\n- Password: Swagger123!\n\nUsa /api/auth/login para obtener el access_token y luego el botón Authorize con el valor Bearer {access_token}.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor actual"
        }
    ],
    "paths": {
...
        "/api/condominiums": {
            "get": {
                "tags": [
                    "Condominios"
                ],
                "summary": "Listar condominios",
                "operationId": "condominiumsIndex",
                "responses": {
                    "200": {
                        "description": "Condominios encontrados"
                    },
                    "403": {
                        "description": "Acceso permitido solo al administrador senior"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Condominios"
                ],
                "summary": "Crear condominio",
                "operationId": "condominiumsStore",
                "requestBody": {
                    "description": "Enviar como multipart/form-data. Debe incluirse address o direction. El logo se almacena en el filesystem configurado para logos.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Condominio Vista Verde"
                                    },
                                    "slug": {
                                        "description": "Identificador único. Si se omite, se genera desde name.",
                                        "type": "string",
                                        "example": "condominio-vista-verde",
                                        "nullable": true
                                    },
                                    "ruc": {
                                        "type": "string",
                                        "example": "0999999999001",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Residencial",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Condominio residencial con áreas comunes y seguridad privada.",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Activo",
                                        "nullable": true,
                                        "enum": [
                                            "Activo",
                                            "Inactivo"
                                        ]
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "administracion@vistaverde.example.com",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+593 4 555 0101",
                                        "nullable": true
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "EC",
                                        "nullable": true
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "example": 9,
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 101,
                                        "nullable": true
                                    },
                                    "address": {
                                        "description": "Dirección del condominio. Obligatoria si no se envía direction.",
                                        "type": "string",
                                        "example": "Av. Principal 123 y Calle Secundaria",
                                        "nullable": true
                                    },
                                    "direction": {
                                        "description": "Alias compatible de address. Obligatoria si no se envía address.",
                                        "type": "string",
                                        "example": "Av. Principal 123 y Calle Secundaria",
                                        "nullable": true
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "Frente al parque central",
                                        "nullable": true
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -2.170998,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -79.922359,
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD",
                                        "nullable": true
                                    },
                                    "towers": {
                                        "type": "integer",
                                        "example": 4,
                                        "nullable": true
                                    },
                                    "houses": {
                                        "type": "integer",
                                        "example": 120,
                                        "nullable": true
                                    },
                                    "total_units": {
                                        "type": "integer",
                                        "example": 120,
                                        "nullable": true,
                                        "minimum": 0
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    },
                                    "characteristics": {
                                        "description": "IDs de items activos del catálogo condominium_features.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ],
                                        "nullable": true
                                    },
                                    "admin_name": {
                                        "type": "string",
                                        "example": "Carlos",
                                        "nullable": true
                                    },
                                    "admin_last_name": {
                                        "type": "string",
                                        "example": "Ramírez",
                                        "nullable": true
                                    },
                                    "admin_document_type": {
                                        "type": "string",
                                        "example": "Cédula",
                                        "nullable": true
                                    },
                                    "admin_id_number": {
                                        "type": "string",
                                        "example": "0912345678",
                                        "nullable": true
                                    },
                                    "admin_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.ramirez@example.com",
                                        "nullable": true
                                    },
                                    "admin_phone": {
                                        "type": "string",
                                        "example": "+593 99 123 4567",
                                        "nullable": true
                                    },
                                    "admin_status": {
                                        "type": "string",
                                        "example": "Activo",
                                        "nullable": true,
                                        "enum": [
                                            "Activo",
                                            "Inactivo"
                                        ]
                                    },
                                    "logo": {
                                        "description": "Imagen del logo. Tamaño máximo: 5 MB.",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Condominio creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno, por ejemplo cuando no es posible almacenar el logo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/condominiums/options": {
            "get": {
                "tags": [
                    "Condominios"
                ],
                "summary": "Listar opciones de condominios para combos",
                "operationId": "condominiumsOptions",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filtrar por nombre del condominio.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        },
                        "example": "Cedros"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opciones de condominios encontradas"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso permitido solo al administrador senior"
                    },
                    "422": {
                        "description": "Parámetros inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/condominiums/{condominium}/status": {
            "patch": {
                "tags": [
                    "Condominios"
                ],
                "summary": "Activar o inactivar condominio",
                "operationId": "condominiumsUpdateStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_active"
                                ],
                                "properties": {
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiToken": {
                "type": "http",
                "bearerFormat": "API token",
                "scheme": "bearer"
            }
        }
    },
...
}